Fill 0x80028564: Skeleton.c fn_80028564 exact (35/52)

+++ src/char.c
@@ -34,7 +34,7 @@ void  fn_8001DC64(Character* pChar, SkinChoices* pChoices);
 void  ClipBank_Restore(int nSlot);                  // skalib.c
 ClipBank* ClipBank_Get(u32 nSlot);                  // skalib.c
 AnimLib* AnimLib_Load(u8* pData, ClipBank* pBank);  // skalib.c
-CharModel* fn_80028564(u8* pData, int n, CharModelDefs* pDefs, u8 b);   // Skeleton.c
+CharModel* fn_80028564(u8* pData, s8 n, CharModelDefs* pDefs, u8 b);    // Skeleton.c
 Skin* fn_800377FC(u8* pData, u8 b);                 // Skin.c
 s32   fn_80037708(void* pSkin);                     // Skin.c
 void  fn_800375AC(Skin* pSkin, int n);              // Skin.c
